+++ example/lib/main.dart
@@ -1,22 +1,11 @@
 import 'package:flutter/material.dart';
-
 import 'package:facebook_analytics/facebook_analytics.dart';
 
 final facebookAnalytics = FacebookAnalytics();
 
 void main() => runApp(MyApp());
 
-class MyApp extends StatefulWidget {
-  @override
-  _MyAppState createState() => _MyAppState();
-}
-
-class _MyAppState extends State<MyApp> {
-  @override
-  void initState() {
-    super.initState();
-  }
-
+class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(